chore: migrate states
