+++ libs/common/src/key-management/account-cryptography/default-account-cryptographic-state.service.ts
@@ -1,19 +1,12 @@
 import { Observable } from "rxjs";
 
 import { WrappedAccountCryptographicState } from "@bitwarden/sdk-internal";
-import { CRYPTO_DISK, StateProvider, UserKeyDefinition } from "@bitwarden/state";
+import { StateProvider } from "@bitwarden/state";
 import { UserId } from "@bitwarden/user-core";
 
-import { AccountCryptographicStateService } from "./account-cryptographic-state.service";
+import { ACCOUNT_CRYPTOGRAPHIC_STATE } from "../state-definitions";
 
-export const ACCOUNT_CRYPTOGRAPHIC_STATE = new UserKeyDefinition<WrappedAccountCryptographicState>(
-  CRYPTO_DISK,
-  "accountCryptographicState",
-  {
-    deserializer: (obj) => obj as WrappedAccountCryptographicState,
-    clearOn: ["logout"],
-  },
-);
+import { AccountCryptographicStateService } from "./account-cryptographic-state.service";
 
 export class DefaultAccountCryptographicStateService implements AccountCryptographicStateService {
   constructor(protected stateProvider: StateProvider) {}
