+++ libs/common/src/key-management/ephemeral-pin-envelope-mapper.ts
@@ -3,7 +3,7 @@ import { EphemeralPinEnvelopeState } from "@bitwarden/sdk-internal";
 import { SdkRecordMapper } from "../platform/services/sdk/client-managed-state";
 import { UserKeyDefinition } from "../platform/state";
 
-import { PIN_PROTECTED_USER_KEY_ENVELOPE_EPHEMERAL } from "./pin/pin.state";
+import { PIN_PROTECTED_USER_KEY_ENVELOPE_EPHEMERAL } from "./state-definitions";
 
 export class EphemeralPinEnvelopeMapper implements SdkRecordMapper<
   EphemeralPinEnvelopeState,
