+++ libs/common/src/key-management/master-password/services/master-password.service.ts
@@ -18,7 +18,6 @@ import { USER_SERVER_CONFIG } from "../../../platform/services/config/default-co
 import {
   MASTER_PASSWORD_DISK,
   MASTER_PASSWORD_MEMORY,
-  MASTER_PASSWORD_UNLOCK_DISK,
   StateProvider,
   UserKeyDefinition,
 } from "../../../platform/state";
@@ -28,6 +27,7 @@ import { KeyGenerationService } from "../../crypto";
 import { CryptoFunctionService } from "../../crypto/abstractions/crypto-function.service";
 import { EncryptedString, EncString } from "../../crypto/models/enc-string";
 import { USES_KEY_CONNECTOR } from "../../key-connector/services/key-connector.service";
+import { MASTER_PASSWORD_UNLOCK_DATA } from "../../state-definitions";
 import { InternalMasterPasswordServiceAbstraction } from "../abstractions/master-password.service.abstraction";
 import {
   MasterKeyWrappedUserKey,
@@ -59,16 +59,6 @@ export const FORCE_SET_PASSWORD_REASON = new UserKeyDefinition<ForceSetPasswordR
   "forceSetPasswordReason",
   {
     deserializer: (reason) => reason,
-    clearOn: ["logout"],
-  },
-);
-
-/** Disk to persist through lock */
-export const MASTER_PASSWORD_UNLOCK_KEY = new UserKeyDefinition<MasterPasswordUnlockData>(
-  MASTER_PASSWORD_UNLOCK_DISK,
-  "masterPasswordUnlockKey",
-  {
-    deserializer: (obj) => MasterPasswordUnlockData.fromJSON(obj),
     clearOn: ["logout"],
   },
 );
@@ -338,20 +328,20 @@ export class MasterPasswordService implements InternalMasterPasswordServiceAbstr
     assertNonNullish(userId, "userId");
 
     await this.stateProvider
-      .getUser(userId, MASTER_PASSWORD_UNLOCK_KEY)
+      .getUser(userId, MASTER_PASSWORD_UNLOCK_DATA)
       .update(() => masterPasswordUnlockData.toJSON());
   }
 
   async clearMasterPasswordUnlockData(userId: UserId): Promise<void> {
     assertNonNullish(userId, "userId");
 
-    await this.stateProvider.getUser(userId, MASTER_PASSWORD_UNLOCK_KEY).update(() => null);
+    await this.stateProvider.getUser(userId, MASTER_PASSWORD_UNLOCK_DATA).update(() => null);
   }
 
   masterPasswordUnlockData$(userId: UserId): Observable<MasterPasswordUnlockData | null> {
     assertNonNullish(userId, "userId");
 
-    return this.stateProvider.getUser(userId, MASTER_PASSWORD_UNLOCK_KEY).state$;
+    return this.stateProvider.getUser(userId, MASTER_PASSWORD_UNLOCK_DATA).state$;
   }
 
   async setLegacyMasterKeyFromUnlockData(
