+++ libs/common/src/platform/services/key-state/user-key.state.ts
@@ -1,6 +1,4 @@
-import { UserKey } from "../../../types/key";
-import { SymmetricCryptoKey } from "../../models/domain/symmetric-crypto-key";
-import { CRYPTO_DISK, CRYPTO_MEMORY, UserKeyDefinition } from "../../state";
+import { CRYPTO_DISK, UserKeyDefinition } from "../../state";
 
 /**
  * Ever had user key is a hack that allows differentiating TDE users that
@@ -21,10 +19,3 @@ export const USER_EVER_HAD_USER_KEY = new UserKeyDefinition<boolean>(
     clearOn: ["logout"],
   },
 );
-
-export const USER_KEY = UserKeyDefinition.record<UserKey>(CRYPTO_MEMORY, "userKey", {
-  deserializer: (obj) => SymmetricCryptoKey.fromJSON(obj) as UserKey,
-  clearOn: ["logout", "lock"],
-  // Prevents the state from caching and rxjs observable becoming hot observable.
-  cleanupDelayMs: 0,
-});
