+++ libs/key-management/src/index.ts
@@ -19,7 +19,8 @@ export {
   fromSdkKdfConfig,
 } from "./models/kdf-config";
 export { KdfConfigService } from "./abstractions/kdf-config.service";
-export { DefaultKdfConfigService, KDF_CONFIG } from "./kdf-config.service";
+export { DefaultKdfConfigService } from "./kdf-config.service";
+export { KDF_CONFIG } from "@bitwarden/common/key-management/state-definitions";
 export { KdfType } from "./enums/kdf-type.enum";
 
 export * from "./user-asymmetric-key-regeneration";
