+++ libs/key-management/src/kdf-config.service.ts
@@ -1,28 +1,11 @@
 import { firstValueFrom, Observable } from "rxjs";
-import { Jsonify } from "type-fest/source/jsonify";
 
-import {
-  KDF_CONFIG_DISK,
-  StateProvider,
-  UserKeyDefinition,
-} from "@bitwarden/common/platform/state";
+import { KDF_CONFIG } from "@bitwarden/common/key-management/state-definitions";
+import { StateProvider } from "@bitwarden/common/platform/state";
 import { UserId } from "@bitwarden/common/types/guid";
 
 import { KdfConfigService } from "./abstractions/kdf-config.service";
-import { KdfType } from "./enums/kdf-type.enum";
-import { Argon2KdfConfig, KdfConfig, PBKDF2KdfConfig } from "./models/kdf-config";
-
-export const KDF_CONFIG = new UserKeyDefinition<KdfConfig>(KDF_CONFIG_DISK, "kdfConfig", {
-  deserializer: (kdfConfig: Jsonify<KdfConfig>) => {
-    if (kdfConfig == null) {
-      return null;
-    }
-    return kdfConfig.kdfType === KdfType.PBKDF2_SHA256
-      ? PBKDF2KdfConfig.fromJSON(kdfConfig)
-      : Argon2KdfConfig.fromJSON(kdfConfig);
-  },
-  clearOn: ["logout"],
-});
+import { KdfConfig } from "./models/kdf-config";
 
 export class DefaultKdfConfigService implements KdfConfigService {
   constructor(private stateProvider: StateProvider) {}
